+++ packages/react/src/components/uploader.tsx
@@ -1,6 +1,7 @@
 "use client";
 
-import type { ErrorMessage, FileRouter } from "uploadthing/server";
+import type { ErrorMessage } from "@uploadthing/shared";
+import type { FileRouter } from "uploadthing/types";
 
 import type { UploadthingComponentProps } from "../types";
 import { UploadButton } from "./button";
