+++ packages/solid/src/components/uploader.tsx
@@ -1,4 +1,5 @@
-import type { ErrorMessage, FileRouter } from "uploadthing/server";
+import type { ErrorMessage } from "@uploadthing/shared";
+import type { FileRouter } from "uploadthing/types";
 
 import type { UploadthingComponentProps } from "../types";
