+++ packages/uploadthing/test/__test-helpers.ts
@@ -3,8 +3,8 @@ import { beforeEach, vi } from "vitest";
 import type { FetchEsque } from "@uploadthing/shared";
 
 import { UPLOADTHING_VERSION } from "../src/internal/constants";
-import type { PSPResponse } from "../src/internal/handler";
-import type { ActionType } from "../src/server";
+import type { ActionType } from "../src/internal/types";
+import type { PSPResponse } from "../src/types";
 
 export const fetchMock = vi.fn();
 export const middlewareMock = vi.fn();
